[REFACTOR] Refatorando diretorio dos arquivos #1177
