+++ front/src/modules/proposta/visualizar/components/slPropostaCustosVinculados.vue
