+++ front/src/modules/proposta/visualizar/components/slPropostaDetalhamentoConsolidacao.vue
