+++ front/src/modules/proposta/visualizar/components/slPropostaFontesDeRecursos.vue
